Return APIResponse object for friend and shop and rename friend.ts files to social.ts
