+++ types/general.ts
@@ -1,0 +1,4 @@
+export type APIResponse = {
+  success: boolean;
+  error: string | null;
+};
